backend/requirements.txt
numpy>=1.26.0
scipy>=1.11.0
opencv-python>=4.8.0.0
Pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
werkzeug>=3.0.0
python-dotenv>=1.0.0
imutils>=0.5.4
urllib3<2.0.0
easyocr>=1.7.0
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.1.0
pytesseract>=0.3.10
